IN_PROGRESS - bug 6277: BufferedDataTable should return number of rows as long https://bugs.knime.org/show_bug.cgi?id=6277
Adapted classes in org.knime.base

+++ org.knime.base/src-deprecated/org/knime/base/node/preproc/missingval/MissingValueHandling2TableIterator.java
@@ -73,8 +73,8 @@ class MissingValueHandling2TableIterator extends RowIterator {
     private final RowIterator m_internIt;
     private final MissingValueHandling2Table m_table;
     private final ExecutionMonitor m_exec;
-    private final int m_finalCount;
-    private int m_count;
+    private final long m_finalCount;
+    private long m_count;
     // since we don't know if to skip the next row in the underlying table
     // (ColSetting.METHOD_IGNORE_ROWS), we save the next to return
     private DataRow m_next;
